Migrated object::Id from u26 to usize. Moved id storage from usize to NonZero<usize>. Added child_id convenience function. Removed visible_entries (possibly temporary).

+++ src/lib.rs
@@ -30,7 +30,7 @@ mod prelude {
     }
     pub mod object {
         pub use crate::object::Kind;
-        pub type Id = super::NonZero<u16>;
+        pub type Id = super::NonZero<usize>;
     }
 
     pub type Anchor = (Span, Rise);
